requirements.txt
litserve
piper-tts
fastapi
pydantic
pydub
torch
onnxruntime-gpu